FST 83 - move some files, added properties to caseData and changed to controlled accordion
