+++ src/features/integration/IntegrationConfigurationForm.tsx
@@ -12,7 +12,7 @@ import IFormData from "./types/Form/FormData";
 import CaseInformation from "./components/CaseInformation";
 import IntegrationService from "./service/IntegrationService";
 import {defaultValues} from "./util/DefaultValues";
-import {mapToDto} from "./util/util";
+import {mapToDto} from "./util/MapToDto";
 
 const useStyles = makeStyles((theme: Theme) =>
     createStyles({
@@ -39,6 +39,12 @@ const useStyles = makeStyles((theme: Theme) =>
 
 const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<any>> = () => {
     const classes = useStyles();
+    const [expanded, setExpanded] = React.useState<string | false>(false);
+    const handleChange =
+        (panel: string) => (event: React.SyntheticEvent, isExpanded: boolean) => {
+            setExpanded(isExpanded ? panel : false);
+        };
+
     const {handleSubmit, watch, setValue, formState: {}} = useForm<IFormData>({
         defaultValues: defaultValues
     });
@@ -78,7 +84,7 @@ const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<
             <Box>
                 <Typography variant={"h5"} sx={{mb: 2}}>Integrasjonskonfigurasjon</Typography>
                 <form className={classes.form} onSubmit={onSubmit}>
-                    <Accordion className={classes.accordion} defaultExpanded={true}>
+                    <Accordion expanded={expanded === 'panel1'} onChange={handleChange('panel1')} className={classes.accordion} defaultExpanded={true}>
                         <AccordionSummary  expandIcon={<ExpandMoreIcon />}>
                             <Typography variant={"h6"}>Integrasjonslogikk</Typography>
                         </AccordionSummary>
@@ -86,7 +92,7 @@ const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<
                             <CaseInformation setValue={setValue} style={classes} caseCreationStrategy={watch("caseData.caseCreationStrategy")} selectedForm={watch("selectedForm")} />
                         </AccordionDetails>
                     </Accordion>
-                    <Accordion className={classes.accordion}>
+                    <Accordion expanded={expanded === 'panel2'} onChange={handleChange('panel2')} className={classes.accordion}>
                         <AccordionSummary expandIcon={<ExpandMoreIcon />}>
                             <Typography variant={"h6"}>Sakspost</Typography>
                         </AccordionSummary>
@@ -94,7 +100,7 @@ const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<
                             <CaseForm style={classes} watch={watch} setValue={setValue}/>
                         </AccordionDetails>
                     </Accordion>
-                    <Accordion className={classes.accordion}>
+                    <Accordion expanded={expanded === 'panel3'} onChange={handleChange('panel3')} className={classes.accordion}>
                         <AccordionSummary expandIcon={<ExpandMoreIcon />}>
                             <Typography variant={"h6"}>Journalpost</Typography>
                         </AccordionSummary>
@@ -102,7 +108,7 @@ const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<
                             <RecordForm style={classes} watch={watch}  setValue={setValue} />
                         </AccordionDetails>
                     </Accordion>
-                    <Accordion className={classes.accordion}>
+                    <Accordion expanded={expanded === 'panel4'} onChange={handleChange('panel4')} className={classes.accordion}>
                         <AccordionSummary expandIcon={<ExpandMoreIcon />}>
                             <Typography variant={"h6"}>Dokument- og objektbeskrivelse</Typography>
                         </AccordionSummary>
@@ -110,7 +116,7 @@ const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<
                             <DocumentForm style={classes} watch={watch} setValue={setValue}/>
                         </AccordionDetails>
                     </Accordion>
-                    <Accordion className={classes.accordion}>
+                    <Accordion expanded={expanded === 'panel5'} onChange={handleChange('panel5')} className={classes.accordion}>
                         <AccordionSummary expandIcon={<ExpandMoreIcon />}>
                             <Typography variant={"h6"}>Avsender</Typography>
                         </AccordionSummary>
@@ -118,7 +124,7 @@ const IntegrationConfigurationForm: React.FunctionComponent<RouteComponentProps<
                             <ApplicantForm style={classes} watch={watch} setValue={setValue}/>
                         </AccordionDetails>
                     </Accordion>
-                    <Accordion className={classes.accordion}>
+                    <Accordion expanded={expanded === 'panel6'} onChange={handleChange('panel6')} className={classes.accordion}>
                         <AccordionSummary expandIcon={<ExpandMoreIcon />}>
                             <Typography variant={"h6"}>Kontroller skjema</Typography>
                         </AccordionSummary>
